[FEAT: #21] - Rota para deletar.
close #21
