+++ app/routers/__init__.py
@@ -1,0 +1,7 @@
+from .burgers import router as router_burger
+from .ingredients_status import router as router_ingredients_status
+
+__all__ = (
+    "router_burger",
+    "router_ingredients_status",
+)
